requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-dotenv

# Whisper Model Requirement
argparse
torch
torchvision
torchaudio
numpy
git+https://github.com/openai/whisper.git

# Antrhopic Claude
anthropic

# LangChain
langchain
langchain_community
chromadb
sentence-transformers